# [refactor][perf]: improve performance when processing incoming chat messages