### In thish PR:
**[refactor][perf]: optimize string replacements when processing incoming chat messages**

- increased performance over the current strategy of creating multiple copies of the entire string and scanning each one.
- about 2x increase in performance for the average message length (see profiling data)

```
Num samples:  4750
Avg Time (overall): 
	[NEW] getRequestMessageCategories:	0.1841ms
	[OLD] getRequestMessageCategories:	0.4019ms

Avg Total Length (chars):  101.31936842105263
Avg ms per Avg Total Length: 
	[NEW] getRequestMessageCategories:	0.1792ms
	[OLD] getRequestMessageCategories:	0.3699ms
```
**[minor][refactor]: move logic from `parseMessageForRequestList` to `getRequestMessageCategories`**

### Images:
![image](https://github.com/user-attachments/assets/18d324aa-1bb6-4cce-8f56-dced4ae45446)